receiptomatic-www: Add approver functionality
Add the ability for an 'approver' to approve or reject receipts.

Approvers will see a new link at the top of the page 'approve receipts'
Clicking this will take them to a new screen showing three receipts per
page that are pending approval.

For each receipt there is one of three choices to be made:

	Skip (default), Approve or Reject.

Skip just means that it won't be processed either way this time around
and will keep showing up in the list until it has been either approved
or rejected.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>