receiptomatic-www: Add a set_proc_title() function
Seeing as Linux doesn't have a setproctitle(3), everyone creates their
own.

Add one based on code from: avahi, linux-util-ng and code found on the
web.

It's a bit hacky and there's a good chance it'll only work on Linux.

The parent process is now called "receiptomatic-www: master" and the
child processes are called "receiptomatic-www: worker"

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>