/receipt_info/: Fix a potential segfault situation
If the tag_info_allowed() check came back with a 0 then would jump to
the end of the function which amongst other things would free the mysql
result and clsoe the mysql connection. However in this case, they would
never have been set up in the first place.

Move the mysql clean up out of and above the out: label.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>