/approve_receipts/: Fix a potential segfault
There may be a case when an incorrect type setting in the passwd table
makes a user an approver but with no valid capabilities. This would then
lead to invalid SQL being generated.

Check that the approver method SQL has been generated or just return
from the function and log an error to the error log.

Also move the GET query check to below this check to avoid any possible
qvars memory leaks.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>