receiptomatic-www: Do checks on entered receipt information
Check the data being submitted from /receipts/. If errors are found the
problematic fields are highlighted in red and should be fixed then this
updated data submitted.

The checks are currently that all fields (except Reason) have something
in them. That the amounts add up including that the vat rate is correct.
Also check that the receipt data is in the right format, isn't in the
future and isn't older than 180 days (MAX_RECEIPT_AGE).

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>