receiptomatic-www: switch to tokyocabinet for the user sessions
Store the user sessions in a tokycabinet table database rather than
sqlite.

I was starting to find the sqlite interface a bit finicky,

I was also seeing some database synchronisation issues which manifested
themselves as failure for the browser to display images sometimes due to
the is_loged_in() check failing due to it not being able to open the
database. The database was likely locked by the session setup stuff in
another process. This could be worked around by retrying the database
access. But see the first problem.

So we switch to use tokyocabinet's table database backend for storing the
user sessions. While it has its own multi-access limitations (multiple
readers can access the table, but a writer will block access, a reader
will also block a writer), in some simple testing it seemed to handle
multiple access more gracefully. I.e if a writer has the file open others
will block until the lock is free'd. It uses fcntl file locking.

And while it takes a bit more code, I find the interface easier to work
with.

In terms of what we store, nothings changed and old sessions are still
cleared out after 4 hours.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>