receiptomatic: Allow users to edit tagged data
Add the ability for users to edit their receipt data, but only
for receipts that are in the PENDING status.

In tag_image() we needed to change the SQL from INSERT INTO to
REPLACE INTO otherwise when editing receipts you'd end up with
multiple entries.

That then required that the id column be made UNIQUE. So make
it UNIQUE in all the tables (which it should be).

In /process_receipts/ we track whether we have come from
/receipt_info/ or not, so we know whether to return back there
or back to /receipts/

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>