cookies: Remove the request_id cookie
Remove the request_id cookie for now, until a working solution is
found.

Essentially the problem is that by the time the browser gets and
stores a new request_id, it's already sent the old one and access
is denied sending the user to the login page.

See docs/BUGS for more details.

I've seen this with multiple browsers and it's also at least been seen
with PHP.

http://uk3.php.net/manual/en/function.session-regenerate-id.php#84242

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>