Fix a segfault in environment parsing
When we get a request from a client, we get various environment
variables such as the query string, the client's ip etc. If any of these
came through as NULL we would segfault when trying to store them to a
local variable.

Check that they are not NULL first, if any of them are NULL we just
return out and go back into accept().

We don't return out if HTTP_COOKIE is NULL as the client might not have
a cookie yet to send.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>