Switch to a runtime config system
Instead of various settings being set at compile time. Switch to a
runtime configuration based system.

This is done through the use of a config file which is read at runtime.
This file is in a simple name=value based format. e.g

	IMAGE_PATH=/path/to/receipt_images
	BASE_URL=https://receiptomatic.example.com
	SESSION_DB=/dev/shm/receiptomatic-www-sessions.tct
	DB_USER=receiptomatic
	DB_PASS=password
	DB_NAME=receiptomatic

For the full list of options, look in www/receiptomatic_config.h

When starting receiptomatic this file should be passed to it as argv[1]

e.g

spawn-fcgi -a 127.0.0.1 -p 9191 -P /tmp/receiptomatic-www.pid --
/path/to/receiptomatic/www/receiptomatic-www /tmp/receiptomatic.cfg

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>