receiptomatic-www: Create and store a user header banner
The user header banner is the bit at the top of the page that displays
the users name, uid and whether they are an approver and or admin.

This is currently created in each html page / function.

Change this to be created at log in time and stored in current_session
(but not in the tokyocabinet database).

This will reduce a lot of duplicated code both in the C and html files,
where this is currently being generated for each page.

What was several lines of code now becomes a simple

ml = TMPL_add_var(ml, "user_hdr", current_session->user_hdr, NULL);

This is mirrored in the html templates.

The www/templates/receipts.tmpl file has been updated to use this.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>