receiptomatic-www: Add the start of user admin functionality
If the user is an admin, they can currently do two things:

    1) List users in the system
    2) Add a user to the system.

When a user has been added, the user is sent an activation email with a
link they need to follow to complete their account set up by setting a
password.

The activation key is valid for 24hrs. If the user attempts to activate
after that, they can generate a new key which is emailed to them.

With the new admin routes, the route handling has changed from using
strstr() to strncmp(), to avoid such problems as /admin/ matching
against /admin/add_user/

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>