Fix compilation breakage in receiver/receiptomatic.c
Commit e3d3115c 'receiptomatic-www: Log failed database connections'
made receiptomatic fail to compile due to including fcgi_stdio.h in
common.h which is in included in db.h which is included in
receiptomatic.c

receiptomatic has no need for the fcgi stuff. Simple fix is to only
include fcgi_stdio.h in common.h if we are building receiptomatic-www
and then only log failed db connections under the same circumstances.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>