Remove the use of the upload email address
Instead of each user having a unique email address to send images to.
There will be just a single address for this. The "From: " address will
now be used to determine who the image is for.

This simplifies things somewhat as there is no longer a need to add an
entry to /etc/aliases (and running of newaliases) for every user added
to the system.

What this now means is that users will need to make sure to send the
images from the email address they use to log into the system.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>