receiptmatic-www: Centralise the setting of the POST/GET variables
Instead of doing this in each function that needs it. Set them near the
beginning of the request and then pass the GHashTable to each function
that needs it rather than the query string.

We don't do this for /process_receipt_approval/ though, as it currently
needs the data in a different form to handle POST array variables.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>