receiptomatic-www: Rework custom field names
As the first steps in moving to using the uid rather than the username
to link users between tables. This required some changes in how the
custom field names were set.

Before, we were using mysql_fetch_row() and then row[n]. Due to adding a
uid field at the start of the row, this pushed all the fields out by
one. Thus it seemed like a good time to switch over to using get_dbrow()
which then led to other re-workings.

Instead of setting the default field names in a structure, we set them
as #define's and then merge these in with the results from get_dbrow()
when needed into the fields structure.

This also updates the db schema which includes the new uid field in
field_names.

That meant we needed to add a '' to the 'INSERT INTO passwd' query in
do_add_user() to account for the added d_reason field.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>