Use uid rather than username to link accross tables
Use the users UID rather than their username in SQL queries that want to
indentify a user.

This is partly motivated by the fact that the users username is their
email address and these can change.

I've not yet decided in the case of a username change, whether to
retroactively update all the old username entries to the new one.

Using the uid alleviates the need for this. It also provides a slight
code reduction.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>