receiptomatic-www: Re-work signal handlers
The signal handlers for dump_session_state and clear_old_sessions where
not async-signal-safe as they used non-reentrant functions such as
printf(), snprintf() and fflush().

There are a couple of things we could do here. One is the self-pipe
trick. p. 1370, The Linux Programming Interface - Michael Kerrisk

This involves the signal handler writing a byte onto a pipe which can be
detected with select/poll etc. This would be needed if for example your
main loop is already waiting on file descriptors, you'd just add the
read end of the pipe to the list of watched file descriptors.

The other way is to have the signal handler set a flag that can be
checked periodically for what actions should be carried out. This is the
method I've gone for.

We sleep in the main loop. We can sleep for a long time (60 seconds in
this case) as the sleep is interrupted by the signal, then the rest of
the loop is executed, where we check the various signal handler flags
for what we need to do and then go back to sleep again.

We also use this new mechanism for the reaping of child pids and the
creating of new server processes if required.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>