utils: Fix a memory leak in get_vars
We need to free the memory returned by url_decode

If we didn't call url_decode, due to not having a value, set it to NULL
so the subsequent free won't cause a segfault.

This has the added benefit that empty values are set to NULL and show up
as such, (null), in the debug log as opposed to just showing some
whitespace.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>