receiptomatic: fix compilation
Commit 47bba0b3 'Update the Makefile' re-worked how _RECEIPTOMATIC_WWW_
was done. Previously it wasn't really having any effect. Now each .o is
compiled with it.

This meant we got lots of FCGI_ errors. Re-work receiptomatic's Makefile
to compile the db.c and get_config.c files from receiptomatic-www
without that define present.

We also need to include a dummy declaration for log_dir, seeing as it's
declared extern in receiptomatic_config.h

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>