receiptomatic-www: Adjust compiler/linker flags
Compile with the following extra flags. (These are standard on Fedora)

-g
    Enable debug information

-D_FORTIFY_SOURCE=2
    Enable compile time buffer checks

-fexceptions
    Enable exceptions handling

-fstack-protector
    Emits extra code to check for buffer overflows

--param=ssp-buffer-size=4
    The minimum size of buffers that will receive stack smashing protection

Also compile with, -PIE to generate Position Independent Executables

Link with full RELRO and pie

Full RELRO causes all symbols to be resolved at program startup which
allows for more things to be made read-only.

This does incur a performance penalty, mainly in start up time. Not
that I've actually noticed and this is a long running application
anyway.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>