receipt_info: Fix trans date display when editing
Commit: c0af816d 'receipt_info: Display the approval/rejection
date/time' broke the displaying of the 'trans date' field when
editing the receipt.

When into edit mode, the trans date field would be displayed as
1970-01-01. This was due to the secs variable being set to the date of
the approval/rejection, however when the reciept is still PENDING, this
is 0.

We only want to set the approval/rejection date/time/reason when the
receipt is either in APPROVED or REJECTED status.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>