receiptomatic-www: Add a simple anti-XSS mechanism
XSS being "Cross Site Scripting".

http://en.wikipedia.org/wiki/Cross-site_scripting
https://www.owasp.org/index.php/XSS

Basically if we're outputing untrusted data, say from a user that has
been entered into a database and is now being sent to a browser for
display. Then we need to take some steps to make sure that data doesn't
include, or has at least been mangled, anything that the browser might
try and parse as html, xml, javascript etc.

libctemplate includes a mechanism for formatting output. We can create a
format function that takes things like <, >, & etc and replaces them with
their hex or textual codes that the browser won't parse, but will still
show as the actual character on the screen.

This commit just adds a de_xss() function. We don't use
TMPL_encode_entity from libctemplate as we encode some different things
and it saves messing with the external library.

Further commits will actually make use of this.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>