receiptomatic-www: Add some anti-csrf infrastructure
Add some infrastructure to help negate CSRF (Cross Site Request Forgery)
attacks.

This introduces a generate_csrf_token() function that when called will
generare an id, similar to the session_id that can be used on forms to
check that the form is being sent from the right person.

You can set this id as a hidden value on the form and when the form is
posted you can check for this value.

See http://en.wikipedia.org/wiki/Csrf for more info on this type of
attack.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>