url_handlers: Fix CSRF issue in /process_receipt/
Commit 4ca2dcb 'CSRF: Protect /receipts/ and /receipt_info/' introduced
a csrf check in process_receipt. This was fine as long as the data
submitted didn't contain errors. If it did have errors that needed the
user to fix, then the resulting /process_receipt/ page did not contain
a csrf token which then caused the subsequent call to process_receipt to
bail out on the csrf check.

To fix, generate a csrf token for the /process_receipt/ page itself.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>