Re-license under the GNU Affero General Public License V3
Previously this was under version 3 of the GNU General Public License.
However, due to this ostensibly being a server side application, people
hosting this and making modifications would not be obliged to provide
said modifications back to their users.

The GNU Affero General Public License was designed to close that
loophole.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>