receiptomatic: Fix compilation failure
Commit deae63ef 'Fix compiler warnings under F16/GCC 4.6.2'. Caused
receiptomatic, the email receiver part, to fail to build due to the
FCGI_printf function being defined for receiptomatic through its
inclusion of db.h, which it couldn't find due to not including
fcgi_stdio.h and isn't what we'd want anyway.

So instead of unconditionally undefining printf and redefining it
again in db.h, we only do that when compiling receiptomatic-www.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>