receiptomatic-www: Initialise the logs first
Do the log file initialisation before trying to load the config file,
thus if we can't open the config file, the error log is already good to
go to log that fact.

This also changes the main program exit to EXIT_FAILURE, as we shouldn't
run all the way through to it.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>