url_handlers: Adjust activations check in admin_pending_activations()
Commit: 290c3e2 'url_handlers: Convert admin_pending_activations() to
new pagination code' introduced a check to see if there really are any
pending activations, due to at least one row being returned from the
database regardless. This checked to see if one of the columns was
empty.

The more correct way is to check the value of 'nrows', if this is 0,
there aren't actually any pending activations.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>