receiptomatic-www: Add a users receipts stats page
Add a page at /stats/ accessible from the receipts landing page, that
displays various stats about a users receipts.

We currently have the following for each of; approved, rejected and
pending. The currency, the number of receipts and receipts gross total
for that currency.

Also we say how many untagged receipts the user has.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>