url_handlers: Add an /admin/stats/ to show overall receipt stats
In order to avoid code duplication of the gathering of receipt stats,
this required a modification to gather_receipt_stats_for_user() to allow
passing -1 for the uid (to signify we want to cover all users), to cater
for this the uid argument was changed to a long long, this should allow
for the -1 as well as the max unsigned int value for real uid's.

We then change our SQL queries to either include the uid clause or not.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>