receiptomatic-www: url_handlers: Enhance get_image()
Modify get_image() to also serve the full size images.

{full,get}_image() do mostly the same thing, the main difference being
that full_image() presents the image as a download so it can be viewed
along side the site.

Enhance get_image() so that if it is serving a full size image (medium
will not be in the path name) it sends some extra headers to tell the
browser to treat it as a download and not to view it inline.

This also has the added benifit of using the file magic to get the mime
type for the full image and then we add attachment to the
Content-Disposition header to force the browser to treat it as a
download.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>