receiptomatic-www: url_handlers: Fix a thinko in get_image()
In commit: 2644fb7 'receiptomatic-www: url_handlers: Enhance
get_image()' We added a check on the requested image url to see if it
included 'medium' or not. For this we used strstr(), however I had the
arguments the wrong way around, it should be; haystack, needle.

Fixed, although somehow it was doing the right thing anyway despite
sending the Content-Disposition header even for medium sized images to
be viewed in-line.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>