www: Fix __func__ on Fedora 16/MySQL/C99
When setting up receiptomatic under Fedora 16, I noticed that __func__
was always returning <unknown> rather than the function name.

This is due to the inclusion of my_global.h for MySQL, specifically this
part of my_global.h in at least MySQL 5.5.22

    /* Provide __func__ macro definition for platforms that miss it. */
    #if __STDC_VERSION__ < 199901L
    #  if __GNUC__ >= 2
    #    define __func__ __FUNCTION__
    #  else
    #    define __func__ "<unknown>"
    #  endif
    #elif defined(_MSC_VER)
    #  if _MSC_VER < 1300
    #    define __func__ "<unknown>"
    #  else
    #    define __func__ __FUNCTION__
    #  endif
    #elif defined(__BORLANDC__)
    #  define __func__ __FUNC__
    #else
    #  define __func__ "<unknown>"
    #endif

and the fact we are compiling with -std=c99 (__STDC_VERSION__ 199901L),
which causes the above checks to fall through to the last else and
define __func__ to just return <unknown>.

If you don't compile with C99 then __func__ gets defined as __FUNCTION__
and works as expected.

See http://bugs.mysql.com/bug.php?id=65294 for the upstream bug report.

To fix this, we undefine __func__ if it has been defined as it is really
just a variable in glibc/gcc.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>