www: Re-work client IP logging
Previously we were using the HTTP_X_FORWARDED_FOR header to determine
the clients IP address. This is fine as long as we are using a proxy
server.

However if we are talking to the web server directly then
HTTP_X_FORWARDED_FOR is either not sent or is blank. Apart from not
showing the clients IP address in the access log, as soon as a login
attempt was made this would crash the server process handling the
request due to calling either inet_pton() with a blank IP address or
getnameinfo() without the previous inet_pton() having been called.

To fix this we rename env_vars.http_x_forwarded_for to
env_vars.remote_addr, which I think is cleaner.

We then set env_vars.remote_addr to the value in HTTP_X_FORWARDED_FOR if
there is one and it's not blank.

If there is no valid value in HTTP_X_FORWARDED_FOR then we use the value
in REMOTE_ADDR.

So this now caters for access either using a proxy server or talking
directly to it.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>