www: Create a wrapper around mysql_real_escape_string()
The purpose of mysql_real_escape_string() is to make safe any strings
that come from user input or are under outside control etc that will be
used in a mysql query, to help prevent SQL injection attacks. It does
this by escaping various characters like  ", ' and some others. e.g

    'Tom's Garage'

would become

    'Tom\'s Garage'

mysql_real_escape_string can be a little unwieldy to use at times with
its four function arguments. So create a simple wrapper around it,
reducing the number of arguments to two, the mysql connection and the
string to make safe.

The string it returns is the string that is safe to pass to mysql and
must be free'd.

I made this a static inline function rather than a macro due to inline
functions seeming to be prefereable to macros and also because of the
somewhat sensitive nature of it, in making sure we pass safe data to
mysql, I thought we'd maybe benefit from the type checking and general
robustness you get with a function.

While this doesn't reduce the number of lines of code needed, it does
make for a bit less typing and less code split across lines. e.g

    safe_string = alloca(strlen(dangerous_string) * 2 + 1);
    mysql_real_escape_string(conn, safe_string, dangerous_string,
			     strlen(dangerous_string));
    /* use safe_string in mysql query */

would now be written as

    safe_string = make_mysql_safe_string(conn, dangerous_string);
    /* use safe_string in mysql query */
    free(safe_string);

Note: We can't use alloca(3) in an inline function.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>