templates: receipts: Fix form submission breakage
Commit 4ca2dcb 'CSRF: Protect /receipts/ and /receipt_info/', broke the
submission of receipt tag information, due to only putting the csrf
token in the delete image form and not the tag form. Meaning that when
tag information was submitted, there was no accompanying csrf token and
the submission duly failed.

So we now put the csrf token in the tag submission form also.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>