db: Add a wrapper around mysql_real_query()
Where before we would do something like

    snprintf(sql, SQL_MAX, "SELECT ...", somevar);
    d_fprintf(sql_log, "%s\n", sql);
    mysql_real_query(conn, sql, strlen(sql));
    res = mysql_store_result(conn);

We can now do

    res = sql_query(conn, "SELECT ...", somevar);

This involves two parts. A macro and a function to do the heavy lifting.

The macro is what should be used in code, as above. This macro then
calls __sql_query() passing it the mysql connection and the query
string, it also passes the function name of the caller for the sql log.

One micro-optimisation we do in __sql_query() is instead of doing
strlen(sql) in the mysql_real_query() call, use the return value from
vsnprintf() which is the number of bytes written into sql and pass that
to mysql_real_query() e.g

    len = vsnprintf(sql, sizeof(sql), ...);
    ...
    mysql_real_query(conn, sql, len);

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>