www: url_helpers: Fix a segfault in /admin/stats/
Commit 888c7a0 'www: url_helpers: Convert
gather_receipt_stats_for_user() to sql_query()' introduced a not 100%
reproducable segfault in url_helpers.c::gather_receipt_stats_for_user()

This commit converted this function to the sql_query() function.
However, for the queries getting a count of untagged images, we didn't
get the mysql return result that was then trying to be used. It would
have been trying to use a previously free'd result, so not sure how this
could have ever worked.

Anyway, get the result returned by sql_query() and everything is happy
again.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>