www: Re-work the setting of custom field names
Previously we were passing a fields structure around and setting and
freeing the field names in each function that wanted them.

The passing of the fields structure around wasn't really correct and not
really needed. It should be treated like the env_vars and user_session
structures which are global, set one and free'd once.

So this makes that happen for the fields structure. It gets set once in
url_handlers() and free'd once in the same.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>