www: environment: Store the name of the requested host
This is the first part of the multi-tenancy work. We will be using the
url/hostname that the user uses to access the site to tell what
organisation they are with. e.g if a user accessed the site by
otl.ri.opentechlabs.net, then we need to get the left most part of the
host, in this case, otl.

To get this, we need to look in some CGI environment variables. We first
look in HTTP_X_FORWARDED_HOST. This will be the case if the web server
in front of receiptomatic is itself behind another web server. Otherwise
if will be in HTTP_HOST.

Once we have the host we store it in env_vars.host, this will still be
the FQDN, a function will be introduced the get only the left most part.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>