multi-tenancy: Make the email receiver multi-tenancy aware
The main difference here is where it will store the images if running in
multi-tenancy mode. Rather than just IMAGE_PATH/UID/ it will now be
IMAGE_PATH/TENANT/UID/.

In order for it to work out who the tenant is, it uses the To: header
from the email and looks at the left most part of the hostname. Thus if
an email was sent to, receipts@otl.ri.opentechlabs.net, this will tell
it that the tenant is otl.

We also make a copy of the db_conn() function from www/db.c here as the
sharing of code was becomming messy and no longer worth it for a few
lines of code.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>