multi-tenancy: utils: Fix send_activation_mail()
In multi-tenancy mode, the hostname in the link in the account
activation email will be different for each organisation. So instead of
using the value in BASE_URL, we just hardcode the use of https and use
the value in env_vars.host for the hostname to use in the activation
link.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>