multi-tenancy: db: Handle incorrect hostnames / db's
Seeing as the application uses the hostname to determine what database
to use, if someone mistypes their url, the app may then be looking for
a database that doesn't exist.

Catch this and put them to a page that suggests they have mistyped the
url and that they should fix it and try again.

Note: This requires that the receiptomatic database user has the GRANT
privilege, so that it gets the 'unknown database' error message and not
an access denied message. Which is used in the above to send the user to
a helpful url.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>