www: db: Re-work database connection handling
Previously a database connection would be made and destroyed in each
function that it was needed. Some actions would mean multiple of these
events occuring during a client connection.

Instead, open the connection once at client connection time and destroy
it at the end of the connection.

This results in a nice reduction of code and _should_ be slightly
faster with doing just one database connection.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>