www: utils: Avoid writing post data to disk
When we got multipart/form-data we wrote it out to a temporary file on
disk, put in the Content-Type header that gmime needs then loaded it
into gmime.

Avoid this temporary file by reading the POST data into memory then
passing it onto gmime.

Any actual files are still written out to disk and listed in the
u_files list.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>