www: Statically allocate various structure members
Statically allocate various members of the user_session structure.

This also fix's a bug in get_tenant(), we ere doing a sizeof(tenant) in
the snprintf(), but tenant is a pointer and the sizeof() will return the
size of the pointer not the actual size of the storgae.

We now use TENANT_MAX + 1 in there.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>