www: utils: generate_hash: Reduce the amount of entropy used
Various bits of the application want to generate a unique key/id and use
a SHA-256 hash for this purpose. To create this hash, we were taking
1024 bytes of data from /dev/urandom.

However, the kernel's entropy source is not endless (without the
employment of HWRNGs) and 1024 bytes is way excessive for its uses
here. Assuming you get the full 256 possible values of a byte from the
entropy source, 1024 bytes could produce 1x10^2466 unique hashes.

That does seem to be the case, in a test creating a million hashes using
1 byte of entropy, produced 256 unique hashes each with ~3-4000
duplicates.

So, lets reduce the number of bytes used to 8, this can produce 256^8
(1.8x10^19) unique hashes. Which I think suffices for things like one
off activation keys, CSRF tokens and even longer lasting things like
session ids.

Also, due to the way the hashing algorithm works, a minor change in
input will generally lead to a very different hash.

Several runs of creating a million hashes using 8 bytes of entropy
produced no duplicates. It's also worth noting that in this case,
producing duplicate hashes over the lifetime of the application
wouldn't generally be a major issue.

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>