www: url_handlers: Fix a compiler warning in process_receipt_approval
Due to commit 60d71e0 'www: utils: Make get_{avar,var}() return a const
char *' we now get the following compiler warning in
url_handlers.c::process_receipt_approval()

    url_handlers.c: In function ‘process_receipt_approval’:
    url_handlers.c:1261:18: warning: initialization discards ‘const’
    qualifier from pointer target type [enabled by default]

This is due to storing the return value from get_avar() into a char *
rather then const char *, however, we also then modify the variable, to
inform whether to skip processing or not, so can't just assign it to
a const.

Fix this up by creating a bool skip variable to use rather then
modifying the variable and use that to check whether to skip processing
or not.

Now we can assign the get_avar() variable as a const char *

Signed-off-by: Andrew Clayton <andrew@opentechlabs.co.uk>