www/url_handlers.c: Fix password changing checks
When a user is changing their details we check to see if they are
changing their password, we do this by looking at the password fields to
see if they are empty or not. However the logic here was incorrect, we
were checking for non-empty fields like

    password1 && password2

However that would actually allow you to enter anything into one of the
password fields while leaving the other empty and could lead to your
password being set to what you typed, even if it was too short.

The fixed loigc is

    password1 || password2

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>