Revert "receiver: Optimise a do { ... } while read loop"
This reverts commit 75d772f0520a84d5524527b8810c612d8fe95143.

This commit changed the do { .. } while loop condition in
create_image_id() from "bytes_read > 0" to "bytes_read == BUF_SIZE".

This is actually not a good paradigm to follow as there are cases where
the kernel may actually send you less data than you requested and
_still_ have more data to send. This can happen at least with pipes.

So it's best to just keep reading until we get _no_ bytes of data.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>