www: utmp: Log the remote port for the login
This logs the port on the remote end that the connection originated
from. In the case of NAT this will be a port on the NAT device.

The main point of this is to allow some chance of linking a login back
to a specific host behind NAT where the logged port will be a port on
the NAT device which you can use to link back to a host behind the NAT
e.g by looking in /proc/net/ip_conntrack on Linux.

This adds a new 'port' column to the utmp table. You can adjust it with
the following SQL

    ALTER TABLE utmp ADD COLUMN 'port' int(5) UNSIGNED NOT NULL AFTER 'ip';

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>