www: audit: Simplify name lookup code in log_login()
By using getaddrinfo(3) we can significantly reduce the amount of code
needed to resolve an IP address to a hostname.

This allows us to do it in an address family agnostic manner and removes
the need to cast stuff around to struct sockaddr, we can just pass the
relevant bits returned from getaddrinfo straight into getnameinfo.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>