www: utils: Allow generate_hash() to create SHA1 hashes
Restructure the code slightly in generate_hash() to look at the type
argument for the type of hash to create and set up the hashing function
accordingly.

Add a SHA1 define to pass to generate_hash() to signify this type of
hash to create.

SHA1's might be preferable to SHA256's sometimes, due to lower space
requirements, 40 bytes vs 64 and their slighty speedier creation.

Using a single core of an Intel Q9400 running at 2.66GHz, I can create
1000 SHA1's in 10ms and 1000 SHA256's in 13ms.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>