www: audit: Fix IPv6 matching
Commit 99c2be2 ("www: Add IP ACL functions") introduced a function
match_ipv6() that would check if a given IPv6 address was within a given
network.

Unfortunately this function had some issues.

It works fine with standard /64 networks but shows some problems when
dealing with prefix lengths that weren't a multiple of 8.

This commit replaces the tcp wrappers based code with code based on the
ngx_ptocidr() & ngx_http_access_inet6() functions from the nginx web
server.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>