receiver: Start to handle errors
If the g_mkdir_with_parents() call fails, don't blindly try and
continue, use perror() to print the reason why it failed and bail out.

When saving the image as sent in the email check we can save it first
and if not print an error message why and bail out.

Also add checks on the mkdir() calls for the medium and small
directories and perror() on error but continue on, e.g if we get File
exists, that's fine. At least if there is a more serious error we'll
have something useful logged.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>