Switch from libctemplate to libflate
This makes a change to the HTML templating library used.

libflate - FLATE Library (Fast Template) by Fabien MENEMENLIS.

This works similarly to libctemplate. However unlike libctemplate which
has loops and if/then/else conditions, libflate keeps all the logic in
the application. It has really three types of construct; zone, table and
variable.

tables are like loops in libctemplate. zones can be used to show or not
show certain parts of the html, like the if/then/else of libctemplates
and then variables are simply substituted for their value.

For example a variable in libctemplate would be like

    <tmpl_var name = "my_variable" />

in libflate this would be

    ##my_variable##

See docs/libflate.txt for some more details.

This all leads to cleaner template files with slightly more logic in the
application, where it belongs.

In libcetemplate you could specify format functions to be used on
output. One such use was for running output through the de_xss()
function. libflate doesn't have this concept however I have implemented
similar functionality in the fcgx version of libflate where for each
variable added to you can optionally specify a function to run it
through.

libflate comes with a checktpl utility to check a template file for
errors. All the templates pass this check.

One other note of worth is that this provides both a reduction in code
and also a reduction in the size of the templates.

For code; 10 files changed, 675 insertions, 735 deletions
For templates; 23 files changed, 905 insertions, 993 deletions

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>